NavigationRecipeModule class created
